requirements.txt
dbt-core==1.7.0
dbt-snowflake==1.7.0
protobuf==4.25.3
dagster==1.6.8
boto3 >= 1.15.0
pandas==1.3.5
dagster==1.7.0